+++ tests/testEnv.php
@@ -1,3 +1,4 @@
 <?php
 $env['lang'] = null;
 $env['root_path'] = realpath(__DIR__ . '/../app');
+$env['as_develop'] = false;